backend/stock_scraping_service/requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.3.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.12.0
google-cloud-pubsub==2.14.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httplib2==0.21.0
idna==3.4
lxml==4.9.3
mypy-extensions==1.0.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.1.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
rsa==4.9
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
uritemplate==4.1.1
urllib3==1.26.14
webdriver-manager==4.0.1
wsproto==1.2.0
